rooms: create common room floor data module
This moves floor data related functions into their own module in TRX,
and ports the basic Room_TestTriggers across as well.
