+++ src/libtrx/include/libtrx/game/rooms.h
@@ -4,4 +4,5 @@
 #include "rooms/const.h"
 #include "rooms/draw.h"
 #include "rooms/enum.h"
+#include "rooms/floor_data.h"
 #include "rooms/utils.h"
