+++ src/tr1/game/room.h
@@ -8,6 +8,3 @@
 #include <stdint.h>
 
 void Room_GetNewRoom(int32_t x, int32_t y, int32_t z, int16_t room_num);
-
-void Room_TestTriggers(const ITEM *item);
-void Room_TestSectorTrigger(const ITEM *item, const SECTOR *sector);
