+++ src/tr2/game/room.h
@@ -10,6 +10,3 @@ void Room_GetNewRoom(int32_t x, int32_t y, int32_t z, int16_t room_num);
 
 // TODO: poor abstraction
 void Room_InitCinematic(void);
-
-void Room_TestTriggers(const ITEM *item);
-void Room_TestSectorTrigger(const ITEM *item, const SECTOR *sector);
